Add back `emmo.sync_reasoner()` test
Fixed factpluspluswrapper for rdflib < 5.0.0.
The `replace` parameter did not become available through the `bind()`
method until rdflib v5.0.0, however, the functionality is there.
Since the `bind()` method is just a proxy method for the same named
method for the `namespace_manager`, this can be used instead.

Note, it is imperative that Java is installed for this to work.