Update EMMO-python -> EMMOntoPy where applicable
This still cannot be updated for GitHub links and the Dockerfile.
While the first is obvious (the repository still has not been renamed),
the latter is less so. The main reason here is that the Docker
instructions outline how one installs the package in within the Docker
container. And the `setup.py` file instructs to find data files in the
shared folder `EMMO-python`.