Add safety CI job and use vMAJOR version actions
Use `vMAJOR` version tags for appliccable actions.

Add a job to run `safety`.
This may also include running `pylint` in the future, due to `pylint`'s
aversion to running in a virtual environment via `pre-commit` due to
importability.

Comment out the "demo" steps of the `tests` job so that it only runs
`pytest` now - which is essentially equivalent to before.