Fixed loading in windows
The Url was normilsed when loading from catalog. This resulted in wrong path in windows. This has now been commented out, and seems to work fine in both windows and linux. It is not completely removed as it might be that issues arise that we have not been able to identify.