Setting prefix explicitly in excelparser (#470)
* Setting prefix explicitly in excelparser
Prefix can now be set explicitly for imported ontologies in excelparser.
Either only prefix of given ontology, or set common prefix for all
ontologies with a given root in the base_iri-

* Added support for reading catalog for each imported ontology in excel2onto
Each imported ontology might have its own catalog-file that should be read. 

* Resolve path to imported ontology if it is on the local file system
If the imported ontology is not a web-location the absolute path is now correctly resolved. 
This is needed if the excelparser is called with input not in the same folder.


* Put excel_parser test in own folder
The test for the excelparser is put in a sperate folder so that the imported ontology is completely
separate from other test ontologies.