Added fix for adding annotations that are not strings in ontodoc
Furthermore, reenabled test for ontodoc. This is still aminimal test,
but I think can be gradually expanded.

Lastlt - because a new altLabel was added in the testontology, it was revieled that
the same entity is returned twice in by get_by_label_all in test_prefix. Issue #511 was
made bacause of this.