Support for data, object and annotation properties
Moved adding of concepts to help_function _add_entities, which can be used to add entities fo type owlread2.Thing/ObjectProperty/AnnotationProperty/DataProperty. _add_entities adds annotation properties, but not relations for classes or ranges and domains for properties

Data properties, object properties and annotation properties should be provided in separate sheets called AnnotationProperties, DataProperties, ObjectProperties
If these sheets do not exist in the excel workbook they will not be considered. The same if the sheets exist, but do not have prefLabels.

An extra column for OtherAnnotations has been added and it is now read correctly.

The dictionary with errors is now returned with separate keys for the different properties and classes.