pre-commit updates
Format is used as argument for ontology.save in owlready2 and cannot
be changed.
ontology.save in owlready2 takes file as argument, ontopy is updated
accordingly and keeping filename as optional with deprecation warning.
Same is done for dir to directory as dir is a predefined built-in in
python and should not be redefined.