Allow controling ontology IRI when saving (#700)
# Description
Allow separating the IRI of the ontology from `base_iri` when saving
using rdflib.

Owlready2 doesn't allow to separate the IRI of the
ontology from the `base_iri` (the namespace of the entities in the
ontology). In many ontologies, e.g. EMMO, are these not the same.

Further work: Also assign the ontology IRI when loading an ontology. 

## Type of change
<!-- Put an `x` in the box that applies. -->
- [ ] Bug fix.
- [x] New feature.
- [ ] Documentation update.
- [ ] Test update.

## Checklist
<!-- Put an `x` in the boxes that apply. You can also fill these out
after creating the PR. -->

This checklist can be used as a help for the reviewer.

- [ ] Is the code easy to read and understand?
- [ ] Are comments for humans to read, not computers to disregard?
- [ ] Does a new feature has an accompanying new test (in the CI or unit
testing schemes)?
- [ ] Has the documentation been updated as necessary?
- [ ] Does this close the issue?
- [ ] Is the change limited to the issue?
- [ ] Are errors handled for all outcomes?
- [ ] Does the new feature provide new restrictions on dependencies, and
if so is this documented?

## Comments
<!-- Additional comments here, including clarifications on checklist if
applicable. -->

---------

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>