Handle owlready2:python_names in genated triples in excelparser (#795)
# Description
EMMOntoPy tests fail with owlready0.45 and higher. 
I want to fix one owlready release at a time, since there seem to be
different things that are
incompatible going up to v0.48.

In v 0.45 owlready2 introduced the python names defined with the
predicate
"http://www.lesfleursdunormal.fr/static/_downloads/owlready_ontology.owl#python_name"
in the ontology. These should not be part of the generated ontology and
have therefore been removed
in the ontology generated with the excelparser.