Added figures to generated documentation (#767)
Added figures to generated documentation. If the value of a property
matches a regex for a figure on the web, it will be included in the
documentation.

**Further work**: The figure size should probably be adjusted to not
make the table too large. Probably preferred to only show a
thumbnail-sized version of the figure in the table. The full size could
be shown by clicking. Easier to play with in a following-up PR when this
is applied to EMMO.

---
Also, added prefLabel as anchoring points such that one can use the
browser to easily get documentation of a concept one knows the prefLabel
of.

For example, entering

    https://w3id.org/emmo#Atom

should bring you to the section in the html doc that documents Atom.

---------

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>