Importing rdfs schemas (#814)
# Description

Partially closes #811.
Closes #797 

I is not possible to easily access classes of typ rdfs:Class. This is
needed if working simultaneously with an rdf-schema (as opposed to an
owl ontology). Owlready2 does not search for enties that are not of type
owl:XXX.
A search for rdfs:Class is now added so that when asking for classes in
an ontology also the rdfs:Class'es are returned as Python objects.

This is not as readily fixed with rdfs:Property. As a first fix, such
properties are now returned as a list of strings (instead of a list of
python objcets). Note that the nodes are present in the onotology, but
the rdfs:Poerty is not cast as a python object.

Owlready2 is designed to support owl ontologies (not rdfs). If we really
want to extend the suport for rdfs:Property the best solution is most
likely to cast it to an owl:Annotation_property. If so this should be
done in a next PR .