Emmocheck now does not check prefLabel for any of entities specific prefixes (#852)
prefixed ommitted "owl.", "0.1.", "bibo.", "core.", "terms.", "vann.",
"schema.org",

Also corrected the emmocheck to fail if status returned is not 0.

NB! We do not only check emmo-terms, because it should be possible to
make emmo-compliant ontologies that do not start with w3id.org/emmo.

# Description
<!-- Summary of change, including the issue(s) to be addressed. -->

## Type of change
<!-- Put an `x` in the box that applies. -->
- [x] Bug fix.
- [ ] New feature.
- [ ] Documentation update.
- [x] Test update.

## Checklist
<!-- Put an `x` in the boxes that apply. You can also fill these out
after creating the PR. -->

This checklist can be used as a help for the reviewer.

- [ ] Is the code easy to read and understand?
- [ ] Are comments for humans to read, not computers to disregard?
- [ ] Does a new feature has an accompanying new test (in the CI or unit
testing schemes)?
- [ ] Has the documentation been updated as necessary?
- [ ] Does this close the issue?
- [ ] Is the change limited to the issue?
- [ ] Are errors handled for all outcomes?
- [ ] Does the new feature provide new restrictions on dependencies, and
if so is this documented?

## Comments
<!-- Additional comments here, including clarifications on checklist if
applicable. -->