Flb/minimal installation (#873)
This pull request introduces improved modularization and dependency
management for optional features (Excel, Graph, Reasoner), refactors
exception handling, and updates documentation and CI workflows
accordingly. The most significant changes are the extraction of optional
dependencies into separate requirement files and extras, the
centralization of exceptions, and the addition of utility functions for
dependency checks.

The extras and minimal version are tested separately in the ci_workflow.

<!-- Summary of change, including the issue(s) to be addressed. -->

## Type of change
<!-- Put an `x` in the box that applies. -->
- [ ] Bug fix.
- [ ] New feature.
- [ ] Documentation update.
- [ ] Test update.

## Checklist
<!-- Put an `x` in the boxes that apply. You can also fill these out
after creating the PR. -->

This checklist can be used as a help for the reviewer.

- [ ] Is the code easy to read and understand?
- [ ] Are comments for humans to read, not computers to disregard?
- [ ] Does a new feature has an accompanying new test (in the CI or unit
testing schemes)?
- [ ] Has the documentation been updated as necessary?
- [ ] Does this close the issue?
- [ ] Is the change limited to the issue?
- [ ] Are errors handled for all outcomes?
- [ ] Does the new feature provide new restrictions on dependencies, and
if so is this documented?

## Comments
<!-- Additional comments here, including clarifications on checklist if
applicable. -->

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: EMMOntoPy Developers <Team4.0@SINTEF.no>